requirements.txt
numpy>=1.24.0
opencv-python==4.13.0.92
PyYAML>=6.0
Pillow>=10.0.0
ultralytics==8.4.14
insightface==0.7.3
mediapipe==0.10.32
easyocr>=1.7.0
torch>=2.1.0
torchvision>=0.16.0
onnxruntime==1.24.1
cleanvision==0.3.7
tqdm==4.67.1
gdown==5.2.1
imagededup==0.3.3.post2
git+https://github.com/openai/CLIP.git